% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{colHpds}
\alias{colHpds}
\title{HPD Intervals}
\usage{
colHpds(parm, conf = 0.95)
}
\arguments{
\item{parm:}{matrix where each column is a sample from a parameter posterior distribution}

\item{conf:}{credibility level}
}
\value{
A matrix where each row is the HPD of each column of the matrix \code{parm}
}
\description{
Computes the High Posterior Density Intervals for each column of a given matrix - requires TeachingDemos package
}
\seealso{
\code{\link[TeachingDemos]{hpd}} and \code{\link[TeachingDemos]{emp.hpd}} in \code{TeachingDemos} package.
}

