% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{dglw}
\alias{dglw}
\title{Density of GLW Finite Mixture}
\usage{
dglw(x, P, Mu, Sigma)
}
\arguments{
\item{x}{quantile}

\item{P}{weights}

\item{Mu}{mean}

\item{S2}{variance}
}
\value{
Density of the Gamma Lognormal Weibull finite mixture with mean Mu, variance S2, and vector of weights P.
}
\description{
\code{dglw} returns the value of the density of the GLW finite mixtue.
}
\seealso{
Other res: \code{\link{pglw}}
}

