% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_single.R
\name{gamma_glw}
\alias{gamma_glw}
\title{Fits the Gamma Model}
\usage{
gamma_glw(parm0, Data, Specs, Specs2 = list(Adaptive = 1000, Periodicity =
  100), prior = NULL)
}
\arguments{
\item{parm0:}{vector of initial values, in the order betas, variance, weights.}

\item{Data:}{list of the data to be analyzed. Must contain: \cr
y: survival times. \cr
d: censorship indicator. The code is 0=right censored, 1=event at time, 2=left censored, 3=interval censored. \cr
X: matrix of covariates (model matrix) or vector of 1's when tere are no covariates.\cr
rint: True or False indicating when to consider different intercept for each component of the finite mixture.}

\item{Specs:}{list of simulation parameters. Must contain: (Same as Laplaces Demon). \cr
Iterations: number of iterations to be simulated. \cr
Status: number of simulations before showing the status. \cr
Thinning: thinning of the generated sample. \cr
Algorithm: algorith used to generate the posterior samples. \cr
LogFile: path to print the simulation status.}

\item{Specs2:}{list of simulation parameters specific to the Algorithm chosen (Same as Laplaces Demon).}

\item{prior:}{list of log of prior distribution for beta an sigma and vector of Dirichlet parameters for the weight. 
Default: list(beta='dnormv(beta, 0, 100, log=T)', sigma='dexp(sigma, 1, log=T)', alpha==c(1,1,1)).}
}
\value{
A LaplacesDemon object which contains also a MCMC sample of the posterior distribution of the GLW parameters.
}
\description{
\code{gamma_glw} generates a MCMC sample considering a gama distribution for the response variable
}

