\name{glw-package}
\alias{glw-package}
\alias{glw}
\docType{package}
\title{
Gama Lognormal Weibull finite mixture
}
\description{
This package fits the Gama Lognormal Weibull finite mixture model.
}
\details{
\tabular{ll}{
Package: \tab glw\cr
Type: \tab Package\cr
Version: \tab 0.0.0.9000\cr
Date: \tab 2016-04-18\cr
License: \tab MIT\cr
}
}
\author{
Brian Melo <brian.rmelo@gmail.com>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
