% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptcr_fit.R
\name{glw_ptcr}
\alias{glw_ptcr}
\title{Fits the Promotion Time Cure Rate model with covariates}
\usage{
glw_ptcr(parm0 = NULL, Data, Specs, Specs2 = list(Adaptive = 1000,
  Periodicity = 100), prior = NULL, mixture = "glw")
}
\arguments{
\item{parm0:}{vector of initial values, in the order betas, variance, weights, cure rate.}

\item{Data:}{list of the data to be analyzed. Must contain: \cr
y: survival times. \cr
d: censorship indicator. The code is 0=right censored, 1=event at time, 2=left censored, 3=interval censored. \cr
X: matrix of covariates (model matrix) or vector of 1's when tere are no covariates.\cr
rint: True or False indicating when to consider different intercept for each component of the finite mixture.}

\item{Specs:}{list of simulation parameters. Must contain: (Same as Laplaces Demon). \cr
Iterations: number of iterations to be simulated. \cr
Status: number of simulations before showing the status. \cr
Thinning: thinning of the generated sample. \cr
Algorithm: algorith used to generate the posterior samples. \cr
LogFile: path to print the simulation status.}

\item{Specs2:}{list of simulation parameters specific to the Algorithm chosen (Same as Laplaces Demon).}

\item{prior:}{list of log of prior distribution for beta an sigma and vector of Dirichlet parameters for the weight. 
Default: list(beta='dnormv(beta, 0, 100, log=T)', sigma='dexp(sigma, 1, log=T)', alpha==c(1,1,1)).}

\item{mixture:}{indicates the finite mixture to be fitted to the data. Options are: 'glw', 'gl', 'gw', 'lw'.}
}
\value{
A LaplacesDemon object which contains also a MCMC sample of the posterior distribution of the GLW parameters.
}
\description{
\code{glw_ptcr} generates a MCMC sample from the Standard Cure Rate model, with covariates on the cure rate and the GLW FM distribution for the susceptible individuals
}

