% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glw_loglik.R
\name{loglik}
\alias{loglik}
\title{Individual Log-Likelihood}
\usage{
loglik(parm, Data, mixture)
}
\arguments{
\item{parm:}{vector of parameters}

\item{Data:}{the same data as in \code{\link{glwfm}}.}

\item{mixture:}{indicates the finite mixture to be fitted to the data. Options are: 'glw', 'gl', 'gw', 'lw'}
}
\value{
Returns a vector with the individual contribution of each observations to the log likelihood
}
\description{
\code{loglik} computes the log likelihood of every observations for a given vector of parameters
}

