% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{pglw}
\alias{pglw}
\title{Distribution Function of GLW Finite Mixture}
\usage{
pglw(x, P, Mu, Sigma, lower.tail = T)
}
\arguments{
\item{x}{vector of quantiles.}

\item{P}{vector of mixture weights.}

\item{Mu}{mean.}

\item{Sigma}{variance.}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].}
}
\value{
CDF of the Gamma Lognormal Weibull finite mixture with mean Mu, variance Sigma, and vector of weights P.
}
\description{
\code{pglw} returns the cdf of the GLW finite mixtue.
}
\seealso{
Other res: \code{\link{dglw}}
}

