% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glw_pred.R
\name{pred}
\alias{pred}
\title{Predictive Distribution of GLW}
\usage{
pred(parm, X, mixture)
}
\arguments{
\item{parm:}{matrix of parameters generated from the posterior distribution.}

\item{X:}{vector containing the new observation covariates.}

\item{mixture:}{indicates the finite mixture to be fitted to the data. Options are: 'glw', 'gl', 'gw', 'lw'.}
}
\value{
A generated sample of the predictive distribution for the new observation.
}
\description{
\code{pred} estimates the predictive distribution for a new observation
}

