% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{rglw}
\alias{rglw}
\title{Random Generation From the GLW Finite Mixture}
\usage{
rglw(n, P, Mu, Sigma)
}
\arguments{
\item{n}{number of observations.}

\item{P}{vector of mixture weights.}

\item{Mu}{mean.}

\item{Sigma}{variance.}
}
\value{
Generates a sample with \code{n} observations from the Gamma Lognormal Weibull finite mixture model with mean Mu, variance Sigma, and vector of weights P.
}
\description{
\code{rglw} generates a sample from the GLW finite mixtue.
}
\seealso{
\code{\link{pglw}} for cdf, \code{\link{dglw}} for density
}

