% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glw_sample.R
\name{rglw2}
\alias{rglw2}
\title{Generate Random Sample}
\usage{
rglw2(n, P, Mu.gamma, Mu.lnorm, Mu.weibull, Sigma)
}
\arguments{
\item{Mu.gamma, Mu.lnorm, Mu.weibull:}{means of the Gamma, Lognormal and Weibull distribution.}

\item{n:}{number of observations.}

\item{P:}{weights of the finite mixture.}

\item{Sigma:}{variance}
}
\description{
Generates a random sample from the GLW finite mixture with different means for the different distributions.
}

