% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glw_sample.R
\name{rglw3}
\alias{rglw3}
\title{Random Generation From The GLW FM Regression Model}
\usage{
rglw3(n, Betas, X, P, Sigma, Pc = c(0, 0))
}
\arguments{
\item{n:}{number of observations.}

\item{Betas:}{regression coefficients.}

\item{X:}{matrix of covariates (with column of 1's).}

\item{P:}{weights of the finite mixture.}

\item{Sigma:}{variance.}

\item{Pc:}{vector with probabiliity of right and left censored times.}
}
\value{
A matrix with first column being the times, the second column the censorship 
indicator and the remaining columns are the covariates matrix \code{X}.
}
\description{
\code{rglw3} generates a sample from the glw finite mixture regression model 
with right OR left censored times given a model matrix and the regression coefficients.
}

