% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glw_sample.R
\name{rscr_glw}
\alias{rscr_glw}
\title{Random Generation From The GLW FM Standard Cure Rate Model}
\usage{
rscr_glw(n, Mu, Sigma, P, Betas, X, pc)
}
\arguments{
\item{n:}{number of observations.}

\item{Mu:}{mean}

\item{Sigma:}{variance.}

\item{P:}{weights of the finite mixture.}

\item{Betas:}{regression coefficients associated with the long term parameter.}

\item{X:}{matrix of covariates (with column of 1's).}

\item{pc:}{vector with probability of right censored times.}
}
\value{
A matrix with first column being the times, the second column the censorship 
indicator and the remaining columns are the covariates matrix \code{X}.
}
\description{
\code{rscr_glw} generates a sample from the glw fm standard cure rate regression model 
with right censored times given a model matrix and the regression coefficients.
}

