% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptcr_survival.R
\name{Spop_ptcr}
\alias{Spop_ptcr}
\title{Estimated Survival Function of GLW Promotion Time Cure Rate Model with covariates}
\usage{
Spop_ptcr(parm, y, X, mixture, by = NULL)
}
\arguments{
\item{parm:}{resulting sample from MCMC simulation.}

\item{y:}{response variable (survival time).}

\item{X:}{vector of covariates for one observation.}

\item{mixture:}{indicates the finite mixture to be fitted to the data. Options are: 'glw', 'gl', 'gw', 'lw'.}

\item{by:}{increment of time.}
}
\value{
A list with two obects: \code{surv} and \code{time}. \code{$surv} contains the estimated survival 
probabilities and \code{$time} the respective time. If \code{by}=NULL then time is equal to y.
}
\description{
\code{S} estimates the survival function of the Standard Cure Rate Model, with covariates, using the GLW FM distribution for the susceptible individuals 
and with vector of covariates \code{X} using a sample from the posterior distribution 
of the GLW parameters.
}

