% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glw_survival.R
\name{surv_glw}
\alias{surv_glw}
\title{Estimated Survival Function of GLW Mixture}
\usage{
surv_glw(parm, y, X, mixture, by = NULL)
}
\arguments{
\item{parm:}{resulting sample from MCMC simulation.}

\item{y:}{response variable (survival time).}

\item{X:}{vector of covariates for one observation.}

\item{mixture:}{indicates the finite mixture to be fitted to the data. Options are: 'glw', 'gl', 'gw', 'lw'.}

\item{by:}{increment of time.}
}
\value{
A list with two obects: \code{surv} and \code{time}. \code{$surv} contains the estimated survival 
probabilities and \code{$time} the respective time. If \code{by}=NULL then time is equal to y.
}
\description{
\code{urv_glw} estimates the survival function of the GLW finite mixture regression model for a 
observation with vector of covariates \code{X} using a sample from the posterior distribution 
of the GLW parameters.
}

